requirements.txt
langchainlangchain-openailangchain-communityopenairequestspython-dotenv